requirements-used.txt
chromadb==1.3.0
dashscope==1.22.2
docling==2.58.0
dspy==2.6.14
langchain==1.0.2
langchain_community==0.4.1
langchain_core==1.0.1
langchain_ollama==1.0.0
langchain_openai==1.0.1
langgraph==1.0.1
langgraph_checkpoint_sqlite==3.0.0
llama_index==0.14.6
numpy==2.3.4
openai==2.6.1
owlready2==0.47
pydantic==2.12.3
pydantic_settings==2.11.0
PyPDF2==3.0.1
pytest==8.3.5
python-dotenv==1.2.1
PyYAML==6.0.3
rank_bm25==0.2.2
Requests==2.32.5
scikit_learn==1.7.2
textual==6.4.0
tiktoken==0.12.0
typing_extensions==4.15.0
